requirements.txt
# Django
Django==5.2
django-admin-sortable==2.3  # For sortable admin interface
django-filter==25.1  # For filtering querysets in views
django-celery-beat==2.8.1  # For handling asynchronous tasks
django-redis==5.4.0  # For caching and session storage
django-silk==5.4.0  # For profiling and inspecting requests

# Django REST Framework
djangorestframework==3.16.0
djangorestframework-link-header-pagination==0.1.1  # For link header pagination

# Others
psycopg2-binary==2.9.10  # PostgreSQL database adapter
coverage==7.9.1
